Changes that go along with fixing messaging
